creado eliminarPieza.html

+++ Desguambios/src/main/java/com/dsg/desguambios/entidades/Imagen.java
@@ -8,7 +8,7 @@ public class Imagen {
 	
 	//@Id
 	//@GeneratedValue(strategy = GenerationType.AUTO)
-	private Integer idImagen;
+	private long idImagen;
